refactor: update training environment parameters and improve code formatting

+++ src/rl/env.py
@@ -10,6 +10,7 @@ from loguru import logger
 from src.config.config_loader import ConfigLoader
 from src.pipeline import PathwayGenerator, CostManager
 
+
 @dataclass
 class GraphObservation:
     x_numerical: np.ndarray  # Shape: (max_departments, numerical_feature_dim)
@@ -28,7 +29,7 @@ class GraphObservation:
             "node_mask": self.node_mask,
             "edge_mask": self.edge_mask,
         }
-    
+
     def __repr__(self) -> str:
         return (
             f"GraphObservation(\n"
@@ -40,6 +41,7 @@ class GraphObservation:
             f" edge_mask shape: {self.edge_mask.shape},\n"
             f")"
         )
+
 
 class LayoutEnv(gym.Env):
     metadata = {"render.modes": ["human"]}
@@ -56,9 +58,12 @@ class LayoutEnv(gym.Env):
 
         # [service_time, service_weight, area, x, y, z, action_flag]
         self.categorical_feature_dim = 1  # [name]
-        self.fixable_features: np.ndarray = np.zeros((self.max_departments, 3), dtype=np.float32)
-        self.moveable_features: np.ndarray = np.zeros((self.max_departments, 3), dtype=np.float32)
-        self.action_flag_feature: np.ndarray = np.zeros((self.max_departments, 1), dtype=np.float32)
+        self.fixable_features: np.ndarray = np.zeros(
+            (self.max_departments, 3), dtype=np.float32)
+        self.moveable_features: np.ndarray = np.zeros(
+            (self.max_departments, 3), dtype=np.float32)
+        self.action_flag_feature: np.ndarray = np.zeros(
+            (self.max_departments, 1), dtype=np.float32)
         self.numerical_feature_dim = (
             self.fixable_features.shape[1] +
             self.moveable_features.shape[1] +
@@ -129,21 +134,25 @@ class LayoutEnv(gym.Env):
         self.fixable_features[:len(fixable_features)] = fixable_features
         self.moveable_features[:len(moveable_features)] = moveable_features
 
-        self.scaler.fit(np.concatenate([self.fixable_features, self.moveable_features, self.action_flag_feature], axis=1))
+        self.scaler.fit(np.concatenate(
+            [self.fixable_features, self.moveable_features, self.action_flag_feature], axis=1))
 
     def _precompute_categorical_features(self):
         slots_name_ids = self.cost_manager.slots_name_id
         self.num_total_travel_node = len(self.cost_manager.travel_times.index)
         self.num_total_slot = len(slots_name_ids)
-        self.index_to_dept_id = {i: name for i, name in enumerate(slots_name_ids)}
-        self.dept_id_to_index = {name: i for i, name in enumerate(slots_name_ids)}
+        self.index_to_dept_id = {i: name for i,
+                                 name in enumerate(slots_name_ids)}
+        self.dept_id_to_index = {name: i for i,
+                                 name in enumerate(slots_name_ids)}
 
     def _compute_numerical_features(self):
         pass
 
     def step(self, action: np.ndarray):
         self.current_step += 1
-        self.logger.info(f"Env {self.env_id} Step {self.current_step}, Action taken: {action}")
+        self.logger.info(
+            f"Env {self.env_id} Step {self.current_step}, Action taken: {action}")
 
         total_reward: float = 0.0
 
@@ -161,7 +170,7 @@ class LayoutEnv(gym.Env):
             info = self._get_info()
             self.logger.warning(f"Invalid action: {action}, reward: {reward}")
             return observation, total_reward + reward, terminated, truncated, info
-        
+
         dept1 = self.index_to_dept_id[idx1]
         dept2 = self.index_to_dept_id[idx2]
 
@@ -174,7 +183,7 @@ class LayoutEnv(gym.Env):
         else:
             self.action_flag_feature[:, 0] = 0.0
         area_cost = self.cost_engine.area_compatibility_cost
-        
+
         step_penalty: float = self.config.constraints.step_penalty
         cost_diff: float = previous_cost - new_cost
         travel_reward: float = cost_diff / (self.initial_cost + 1e-6) * 100.0
@@ -182,8 +191,9 @@ class LayoutEnv(gym.Env):
 
         total_reward += travel_reward + step_penalty + area_cost
 
-        self.logger.info(f"Step {self.current_step}: Swapped {dept1} <-> {dept2}, total_reward: {total_reward}, travel_reward: {travel_reward}, area_cost: {area_cost}")
-        
+        self.logger.info(
+            f"Step {self.current_step}: Swapped {dept1} <-> {dept2}, total_reward: {total_reward}, travel_reward: {travel_reward}, area_cost: {area_cost}")
+
         terminated = False
         truncated = self.current_step >= self.max_step
 
@@ -191,7 +201,6 @@ class LayoutEnv(gym.Env):
         info = self._get_info()
 
         return observation.to_dict(), total_reward, terminated, truncated, info
-
 
     def reset(self, seed: Optional[int] = None) -> tuple[Dict[str, np.ndarray], Dict]:
         super().reset(seed=seed)
@@ -216,15 +225,19 @@ class LayoutEnv(gym.Env):
         return observation.to_dict(), info
 
     def _get_observation(self) -> GraphObservation:
-        x_categorical = np.ones((self.max_departments,), dtype=np.int32) * self.PADDING_IDX
+        x_categorical = np.ones((self.max_departments,),
+                                dtype=np.int32) * self.PADDING_IDX
         node_mask = np.zeros((self.max_departments,), dtype=np.int32)
 
-        shuffled_indexes = [self.dept_id_to_index[slot_id] for slot_id in self.cost_engine.layout.values()]
+        shuffled_indexes = [self.dept_id_to_index[slot_id]
+                            for slot_id in self.cost_engine.layout.values()]
         moveable_feature = np.zeros_like(self.moveable_features)
-        moveable_feature[: self.num_total_slot, :] = self.moveable_features[shuffled_indexes]
-        
+        moveable_feature[: self.num_total_slot,
+                         :] = self.moveable_features[shuffled_indexes]
+
         x_norm_numerical = self.scaler.transform(
-            np.concatenate((self.fixable_features, moveable_feature, self.action_flag_feature), axis=1)
+            np.concatenate((self.fixable_features, moveable_feature,
+                           self.action_flag_feature), axis=1).astype(np.float32)
         )
 
         x_categorical[: self.num_total_slot] = np.array(
@@ -245,7 +258,7 @@ class LayoutEnv(gym.Env):
                 edge_index[1, i] = idx2
                 edge_weight[i] = weight
                 edge_mask[i] = 1
-        
+
         return GraphObservation(
             x_numerical=x_norm_numerical,
             x_categorical=x_categorical,
@@ -263,8 +276,10 @@ class LayoutEnv(gym.Env):
             "num_departments": self.num_total_slot,
         }
 
-    def render(self, mode:str="human"):
+    def render(self, mode: str = "human"):
         if mode == "human":
             print(f"Step: {self.current_step}")
-            print(f"Current Total Travel Cost: {self.current_cost:.2f} (Initial: {self.initial_cost:.2f})")
-            print(f"Improvement: {(self.initial_cost - self.current_cost) / (self.initial_cost + 1e-6) * 100:.2f}%")
+            print(
+                f"Current Total Travel Cost: {self.current_cost:.2f} (Initial: {self.initial_cost:.2f})")
+            print(
+                f"Improvement: {(self.initial_cost - self.current_cost) / (self.initial_cost + 1e-6) * 100:.2f}%")
